Prepare training selector view

+++ src/classes/Variable.java
@@ -5,11 +5,11 @@ public class Variable {
 	int process_id;
 	String var_name;
 	String var_type;
-	float min_value;
-	float max_value;
+	double min_value;
+	double max_value;
 	
 	public Variable(int process_id, String var_name,
-			String var_type, float min_value, float max_value) {
+			String var_type, double min_value, double max_value) {
 		super();
 		this.process_id = process_id;
 		this.var_name = var_name;
@@ -19,7 +19,7 @@ public class Variable {
 	}
 	
 	public Variable(int var_id, int process_id, String var_name,
-			String var_type, float min_value, float max_value) {
+			String var_type, double min_value, double max_value) {
 		super();
 		this.var_id = var_id;
 		this.process_id = process_id;
@@ -53,16 +53,16 @@ public class Variable {
 	public void setVar_type(String var_type) {
 		this.var_type = var_type;
 	}
-	public float getMin_value() {
+	public double getMin_value() {
 		return min_value;
 	}
-	public void setMin_value(float min_value) {
+	public void setMin_value(double min_value) {
 		this.min_value = min_value;
 	}
-	public float getMax_value() {
+	public double getMax_value() {
 		return max_value;
 	}
-	public void setMax_value(float max_value) {
+	public void setMax_value(double max_value) {
 		this.max_value = max_value;
 	}
 }
